src/requirements_cuda12.txt
aiofiles==22.1.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.12.1
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.11.0
bleach==6.1.0
boto3==1.29.6
botocore==1.32.6
cachetools==5.3.2
certifi==2023.11.17
certipy==0.1.3
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
confluent-kafka==2.3.0
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
db-dtypes==1.1.1
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.7.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
fonttools==4.45.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
gcsfs==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-bigquery==3.13.0
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-cloud-monitoring==2.16.0
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
grpcio-status==1.59.3
gyp==0.1
hmsclient-hive-3==0.1.1
httplib2==0.22.0
hvac==2.0.0
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.12.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter-server-mathjax==0.2.6
jupyter-telemetry==0.1.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterhub==4.0.2
jupyterlab==3.6.6
jupyterlab-execute-time==2.3.1
jupyterlab-lsp==4.3.0
jupyterlab-widgets==3.0.9
jupyterlab_code_formatter==2.2.1
jupyterlab_git==0.50.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbdime==4.0.1
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opensearch-py==2.4.2
overrides==7.4.0
packaging==23.2
pamela==1.1.0
pandas==2.0.3
pandas-gbq==0.19.2
pandocfilters==1.5.0
papermill==2.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
proto-plus==1.22.3
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAthena==3.0.10
pycparser==2.21
pycryptodome==3.19.0
pydata-google-auth==1.8.2
PyGithub==2.1.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
redis==5.0.1
referencing==0.31.0
requests==2.31.0
requests-aws4auth==1.2.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.1
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
s3transfer==0.7.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
Send2Trash==1.8.2
six==1.16.0
skt==0.2.99
slackclient==2.9.4
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
ssh-import-id==5.11
stack-data==0.6.3
streamz==0.6.4
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.2.0
thrift==0.16.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
torch==2.1.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.18
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
wrapt==1.16.0
y-py==0.6.2
yarl==1.9.3
ypy-websocket==0.8.4
zict==3.0.0
zipp==3.17.0
